Create Reshape Matrix
In MATLAB, there is a handy function called reshape that reshapes a matrix of dimensions m x n into a new one with a different size r x c keeping its original data in row-traversing order.

You are given an m x n matrix mat and two integers r and c representing the number of rows and columns of the wanted reshaped matrix. The reshaped matrix should be filled with all the elements of the original matrix in the same row-traversing order as they were.

If the reshape operation with the given parameters is possible and legal, output the new reshaped matrix. Otherwise, output the original matrix.



Input Parameters:

mat (List[List[int]]): A 2D list representing an m x n matrix.

r (int): The number of rows for the reshaped matrix.

c (int): The number of columns for the reshaped matrix.